Add google analytics env vars
