+++ src/api/analytics.ts
@@ -1,26 +1,6 @@
-import type { AnalyticsEvent } from '@faststore/sdk'
 import type { GatsbyFunctionRequest, GatsbyFunctionResponse } from 'gatsby'
 
-// https://developers.google.com/analytics/devguides/collection/protocol/ga4/sending-events?client_type=gtag
-const measurementId = ''
-const secretApi = ''
-const ANALYTICS_URL = `https://www.google-analytics.com/mp/collect?measurement_id=${measurementId}&api_secret=${secretApi}`
-
-interface AnalyticsRequestData {
-  client_id: string
-  event: AnalyticsEvent
-}
-const parseEventToGA4 = ({ client_id, event }: AnalyticsRequestData) => ({
-  client_id,
-  events: [
-    {
-      name: event.type,
-      params: {
-        ...event.data,
-      },
-    },
-  ],
-})
+import sendGAEvent from '../server/GoogleAnalytics'
 
 const validateRequest = (
   req: GatsbyFunctionRequest,
@@ -28,13 +8,6 @@ const validateRequest = (
 ) => {
   if (req.method !== 'POST') {
     res.statusCode = 405
-    res.send('')
-
-    return false
-  }
-
-  if (!req.body.client_id) {
-    res.statusCode = 400
     res.send('')
 
     return false
@@ -53,13 +26,7 @@ const handler = async (
 
   const { body } = req
 
-  fetch(ANALYTICS_URL, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(parseEventToGA4(body)),
-  })
+  sendGAEvent(body)
 
   res.setHeader('content-type', 'application/json')
   res.statusCode = 200
